1. 加入场景粒子系统

+++ trunk/code/workspace/000_Engine_G2D/g2d/com/g2d/display/DisplayObject.java
@@ -176,7 +176,7 @@ public abstract class DisplayObject extends DObject implements Comparable<Displa
 //	--------------------------------------------------------------------------------------------------------------------------------------------
 
 	public int compareTo(DisplayObject o) {
-		return (int)(z+priority) - (int)(o.z+o.priority);
+		return (int)((z+priority) - (o.z+o.priority));
 	}
 	
 	/**
